Unused project
ironic and ironic-inspector both use ironic-tempest-plugin, so
ironic-inspector-tempest-plugin is not needed, nor was it ever used.

This removes all the content (the sole .gitreview file) and adds a
README to document the status of the project.

We don't remove the git repo because gerrit doesn't have a project
deletion feature.

This mostly follows the instructions at
https://docs.openstack.org/infra/manual/drivers.html#retiring-a-project.

Change-Id: I0f4562677bde5dbb471fc86f606ff5061a11300f